requirements.txt
annotated-types==0.7.0
anyio==4.10.0
beautifulsoup4==4.13.5
certifi==2025.8.3
charset-normalizer==3.4.3
deprecation==2.1.0
et_xmlfile==2.0.0
greenlet==3.2.4
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
numpy==2.3.3
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
playwright==1.55.0
postgrest==1.1.1
pydantic==2.11.9
pydantic_core==2.33.2
pyee==13.0.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
realtime==2.7.0
requests==2.32.5
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
storage3==0.12.1
StrEnum==0.4.15
supabase==2.18.1
supabase_auth==2.12.3
supabase_functions==0.10.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
